pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "facecraft"
version = "1.0.0"
description = "AI Portrait Processing API - Professional headshot processing with background removal, face detection, and enhancement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "OBVT Toolbox"}
]
keywords = [
    "portrait",
    "headshot",
    "face-detection",
    "background-removal",
    "image-processing",
    "api",
    "fastapi",
    "ai",
    "machine-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0,<2.0",
    "Pillow>=10.0.0",
    "rembg>=2.0.50",
    "onnxruntime>=1.15.0",
    "dlib>=19.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "basicsr>=1.4.2",
    "facexlib>=0.3.0",
    "realesrgan>=0.3.0",
    "lpips>=0.1.4",
    "kornia>=0.7.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/obvt/facecraft"
Documentation = "https://github.com/obvt/facecraft#readme"
Repository = "https://github.com/obvt/facecraft"
Issues = "https://github.com/obvt/facecraft/issues"

[project.scripts]
facecraft = "facecraft.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
